moved contexts to a directory with seperate files
