+++ src/test-utils/mocks/contexts/index.ts
@@ -1,0 +1,12 @@
+// Base context utilities
+export { createBaseMockContextState, mockSetFilter } from './base';
+
+// Categories context mocks
+export { 
+    mockCategoriesContextValue, 
+    mockCategoriesContextValueLoading, 
+    mockCategoriesContextValueEmpty 
+} from './categories';
+
+// MeContext provider and types
+export { MeContextProvider, type MeContextProviderProps } from './me';
